GPII-2666: Made Cheerio.js and Markdown-It optional dependencies.
