+++ package.json
@@ -21,12 +21,11 @@
     "license": "BSD-3-Clause",
     "homepage": "https://github.com/GPII/gpii-diff#readme",
     "dependencies": {
-        "cheerio": "1.0.0-rc.2",
         "gpii-handlebars": "1.1.0-dev.20170927T151006Z.df42535",
-        "infusion": "3.0.0-dev.20171102T191539Z.c0636b395",
-        "markdown-it": "8.4.0"
+        "infusion": "3.0.0-dev.20171102T191539Z.c0636b395"
     },
     "devDependencies": {
+        "cheerio": "1.0.0-rc.2",
         "eslint": "4.9.0",
         "eslint-config-fluid": "1.1.0",
         "eslint-plugin-markdown": "1.0.0-beta.6",
@@ -39,6 +38,7 @@
         "grunt-markdownlint": "1.0.42",
         "grunt-jsonlint": "1.1.0",
         "istanbul": "0.4.5",
+        "markdown-it": "8.4.0",
         "node-jqunit": "1.1.7",
         "rimraf": "2.6.2"
     }
