+++ src/js/diff.js
@@ -11,21 +11,6 @@ var gpii  = fluid.registerNamespace("gpii");
 
 fluid.registerNamespace("gpii.diff");
 
-/*
-
-    Cheerio.js is a node-based jQuery alternative to allow parsing and interrogating HTML.  This stub allows us to use
-    the same `load` method on both node and the browser.
-
-*/
-gpii.diff.cheerioBrowser = {
-    load: function (htmlString) {
-        return $(htmlString);
-    }
-};
-
-var MarkDownIt = typeof require !== "undefined" ? require("markdown-it") : window.markdownit;
-var cheerio    = typeof require !== "undefined" ? require("cheerio") : gpii.diff.cheerioBrowser;
-
 /**
  *
  * Produce a representation of a "single value" change, with no deeper comparison.  Intended for use when either the
@@ -205,43 +190,6 @@ gpii.diff.compareStrings = function (leftString, rightString) {
  */
 gpii.diff.isStringNullOrUndefined = function (value) {
     return typeof value === "string" || value === undefined || value === null;
-};
-
-/**
- *
- * Compare two Markdown strings and report their textual differences.
- *
- * @param leftMarkdown {String} - A string containing markdown.
- * @param rightMarkdown {String} - A string to compare to `leftMarkdown`.
- * @param markdownItOptions {Object} - Configuration options to pass to MarkdownIt.  See their docs for supported options.
- * @returns {*}
- */
-gpii.diff.compareMarkdown = function (leftMarkdown, rightMarkdown, markdownItOptions) {
-    if (leftMarkdown === undefined || rightMarkdown === undefined || typeof leftMarkdown !== "string" || typeof rightMarkdown !== "string") {
-        return gpii.diff.singleValueDiff(leftMarkdown, rightMarkdown);
-    }
-    else {
-        var leftString = gpii.diff.markdownToText(leftMarkdown, markdownItOptions);
-        var rightString = gpii.diff.markdownToText(rightMarkdown, markdownItOptions);
-        return gpii.diff.compareStrings(leftString, rightString);
-    }
-};
-
-/**
- *
- * Render a string containing markdown as HTML, then return the textual content.
- *
- * @param markdown {String} - A string containing markdown.
- * @param markdownItOptions {Object} - Configuration options to pass to MarkdownIt.  See their docs for supported options.
- * @returns {String} - The textual content.
- *
- */
-gpii.diff.markdownToText = function (markdown, markdownItOptions) {
-    var mdRenderer = new MarkDownIt(markdownItOptions);
-    var html = mdRenderer.render(markdown);
-    var $ = cheerio.load(html);
-    // The rendering cycle introduces a trailing carriage return that we explicitly remove.
-    return $.text().replace(/[\r\n]+$/, "");
 };
 
 /**
@@ -614,4 +562,9 @@ gpii.diff.compare = function (leftElement, rightElement, compareStringsAsMarkdow
     else {
         return gpii.diff.singleValueDiff(leftElement, rightElement);
     }
+};
+
+// A "stub" function to fail with a meaningful error if markdown support is called when it has not be properly loaded.
+gpii.diff.compareMarkdown = function () {
+    fluid.fail("You must install the optional required dependencies and call `gpii.diff.loadMarkdownSupport()` to use this function");
 };
