+++ tests/static/tests-compareMarkdown.html
@@ -31,6 +31,7 @@
 
     <!-- Our client dependencies under test -->
     <script type="text/javascript" src="/src/js/diff.js"></script>
+    <script type="text/javascript" src="/src/js/markdown-diff.js"></script>
 </head>
 <body>
 <h1 id="qunit-header">Markdown Comparison Tests</h1>
